make saml and confidential callback route public
